+++ builder/init-1.sh
@@ -17,7 +17,12 @@
 # along with Custom Desktop Builder.  If not, see <https://www.gnu.org/licenses/>.
 
 set -e
-mkdir -p ./build/
+
+if [ ! -e ./build/ubuntu-system-adjustments_*-dummy_all.deb ]; then
+    echo "./build/ubuntu-system-adjustments_*-dummy_all.deb not found"
+    exit 1
+fi
+
 wget -c -P ./build/ http://packages.linuxmint.com/pool/main/l/linuxmint-keyring/linuxmint-keyring_2016.05.26_all.deb
 sudo apt install -y ./build/linuxmint-keyring_*_all.deb
 sudo apt-key del 451BBBF2
@@ -49,16 +54,6 @@ Pin-Priority: 1000
 EOF
 
 sudo apt update
-
-if [ ! -f ./build/ubuntu-system-adjustments_*-dummy_all.deb -o ./build/ubuntu-system-adjustments_*-dummy_all.deb -ot ../ubuntu-system-adjustments ]; then
-    rm -f ./build/ubuntu-system-adjustments_*-dummy_all.deb
-    sudo apt install -y equivs
-    make -C ../ ubuntu-system-adjustments
-    mv ../build/ubuntu-system-adjustments_*-dummy_all.deb ./build/
-    make -C ../ clean
-    sudo apt purge -y autoconf automake autopoint autotools-dev binutils binutils-common binutils-x86-64-linux-gnu build-essential debhelper debugedit dh-autoreconf dh-strip-nondeterminism dpkg-dev dwz equivs fakeroot g++ g++-11 gcc gcc-11 gettext intltool-debian libalgorithm-diff-perl libalgorithm-diff-xs-perl libalgorithm-merge-perl libarchive-cpio-perl libarchive-zip-perl libasan6 libatomic1 libbinutils libc-dev-bin libc-devtools libc6-dev libcc1-0 libcrypt-dev libctf-nobfd0 libctf0 libdebhelper-perl libdpkg-perl libfakeroot libfile-fcntllock-perl libfile-stripnondeterminism-perl libgcc-11-dev libitm1 liblsan0 libltdl-dev libmail-sendmail-perl libnsl-dev libquadmath0 libsigsegv2 libstdc++-11-dev libsub-override-perl libsys-hostname-long-perl libtirpc-dev libtool libtsan0 libubsan1 linux-libc-dev lto-disabled-list m4 make manpages-dev po-debconf rpcsvc-proto
-fi
-
 sudo apt install -y ./build/ubuntu-system-adjustments_*-dummy_all.deb
 sudo apt-mark auto ubuntu-system-adjustments
 sudo apt install -y firefox flatpak gnome-session gnome-software gnome-software-plugin-flatpak gnome-software-plugin-snap-
